Changed on_quit behavoir
Follows the approach suggested in conversation of #68 to get rid of the `on_quit` argument, as we also discussed verbally yesterday.
Updated documentation to reflect this change.
Now, one can pass `optimizer=None` and still save state_dict of the model as well as loss_sum and metadata without saving the optimizer. ´check_user_events()` in still saves players with "on_quit" as part of the filename.

Co-Authored-By: RasmusBrostroem <79468407+RasmusBrostroem@users.noreply.github.com>